Replace variable in "topic" settings (#2)
* Fix test case to work in other timezones than UTC

Since logstash converts the given timestamp to UTC, this part
in the "expected message" depends on the timezone where the test
case is executed. With calculating this, the code is now timezone
independent.

* Add test case to check variable topic settings

In MQTT it makes sense to have parts of the topics depending on
parts of an event, e.g. an identifier of the sender of the message.

* Replace variables in "topic" settings

New feature:

mqtt {
  ...
  topic => "data/%{myfield}"
}